Refactor Review Drawer styles to one scss file
