+++ client/src/assets/styles_unused/RatingFilter.css
@@ -1,0 +1,17 @@
+.rating-filter-container {
+  flex-grow: 3;
+  flex-shrink: 1;
+  margin-top: 35px;
+  max-width: 360px;
+  padding-bottom: 5px;
+}
+
+@media (min-width: 1366px) {
+  .rating-filter-container {
+    max-width: 420px;
+  }
+}
+
+fieldset div:first-of-type {
+    margin-top: 0px;
+}
