+++ client/src/assets/styles_unused/RatingSummary.css
@@ -1,0 +1,217 @@
+.customer-rating-heading {
+    font-size: 17px;
+    padding-bottom: 10px;
+}
+
+.overall-rating {
+    display: inline-block;
+    text-align: start;
+    font-size: 40px;
+    vertical-align: top;    
+}
+
+.stars-and-ratings {
+    display: grid;
+    grid-template-rows: 1fr 21px;
+    /* flex-wrap: wrap;
+    flex-grow: 1;
+    flex-basis: 135px; */
+    margin-top: 5px;
+}
+
+.total-reviews {
+    /* flex-shrink: 0; */
+    margin-top: 5px;
+}
+
+.all-reviews-link {
+    display: block;
+    margin-top: 15px; 
+}
+
+.recommend-pct {
+    text-align: left;
+    margin-top: 10px;
+}
+
+.stars {
+    background-image: url(../../assets/bby-stars.png);
+    background-repeat: no-repeat;
+    display: inline-block;
+    position: relative;
+    margin-bottom: 2px;
+}
+
+.stars-medium {
+    width: 137px;
+    height: 26px;
+}
+
+/*
+medium row height: 34px
+medium star width: 27px
+*/
+
+.stars-medium-0-0 {
+    background-position: -100px -606px;
+}
+
+.stars-medium-5-0 {
+    background-position: -100px -336px;
+}
+
+.stars-medium-4-9 {
+    background-position: -100px -370px;
+}
+
+.stars-medium-4-8, .stars-medium-4-7 {
+    background-position: -100px -404px;
+}
+
+.stars-medium-4-6 {
+    background-position: -100px -438px;
+}
+
+.stars-medium-4-5 {
+    background-position: -100px -472px;
+}
+
+.stars-medium-4-4 {
+    background-position: -100px -504px;
+}
+
+.stars-medium-4-3, .stars-medium-4-2 {
+    background-position: -100px -538px;
+}
+
+.stars-medium-4-1 {
+    background-position: -100px -572px;
+}
+
+.stars-medium-4-0 {
+    background-position: -127px -336px;
+}
+
+.stars-medium-3-9 {
+    background-position: -127px -370px;
+}
+
+.stars-medium-3-7, .stars-medium-3-8 {
+    background-position: -127px -404px;
+}
+
+.stars-medium-3-6 {
+    background-position: -127px -438px;
+}
+
+.stars-medium-3-5 {
+    background-position: -127px -472px;
+}
+
+.stars-medium-3-4 {
+    background-position: -127px -504px;
+}
+
+.stars-medium-3-3, .stars-medium-3-2 {
+    background-position: -127px -538px;
+}
+
+.stars-medium-3-1 {
+    background-position: -127px -572px;
+}
+
+.stars-medium-3-0 {
+    background-position: -154px -336px;
+}
+
+.stars-medium-2-9 {
+    background-position: -154px -370px;
+}
+
+.stars-medium-2-7, .stars-medium-2-8 {
+    background-position: -154px -404px;
+}
+
+.stars-medium-2-6 {
+    background-position: -154px -438px;
+}
+
+.stars-medium-2-5 {
+    background-position: -154px -472px;
+}
+
+.stars-medium-2-4 {
+    background-position: -154px -504px;
+}
+
+.stars-medium-2-3, .stars-medium-2-2 {
+    background-position: -154px -538px;
+}
+
+.stars-medium-2-1 {
+    background-position: -154px -572px;
+}
+
+.stars-medium-2-0 {
+    background-position: -181px -336px;
+}
+
+.stars-medium-1-9 {
+    background-position: -181px -370px;
+}
+
+.stars-medium-1-7, .stars-medium-1-8 {
+    background-position: -181px -404px;
+}
+
+.stars-medium-1-6 {
+    background-position: -181px -438px;
+}
+
+.stars-medium-1-5 {
+    background-position: -181px -472px;
+}
+
+.stars-medium-1-4 {
+    background-position: -181px -504px;
+}
+
+.stars-medium-1-3, .stars-medium-1-2 {
+    background-position: -181px -538px;
+}
+
+.stars-medium-1-1 {
+    background-position: -181px -572px;
+}
+
+.stars-medium-1-0 {
+    background-position: -208px -336px;
+}
+
+.stars-medium-0-9 {
+    background-position: -208px -370px;
+}
+
+.stars-medium-0-7, .stars-medium-0-8 {
+    background-position: -208px -404px;
+}
+
+.stars-medium-0-6 {
+    background-position: -208px -438px;
+}
+
+.stars-medium-0-5 {
+    background-position: -208px -472px;
+}
+
+.stars-medium-0-4 {
+    background-position: -208px -504px;
+}
+
+.stars-medium-0-3, .stars-medium-0-2 {
+    background-position: -208px -538px;
+}
+
+.stars-medium-0-1 {
+    background-position: -208px -572px;
+}
