+++ client/src/assets/styles_unused/ReviewFilters.css
@@ -1,0 +1,6 @@
+.review-filters-container {
+    display: flex;
+    position: relative;
+    border-top: 1px solid #c5cbd5;
+    padding: 25px 0;
+}
