+++ client/src/assets/styles_unused/ReviewItem.css
@@ -1,0 +1,152 @@
+li.review-item {
+    list-style: none;
+    display: flex;
+    padding: 30px 0;
+    border-bottom: 1px solid #c5cbd5;
+}
+
+h4 {
+    padding: 0;
+}
+
+.review-item-header {
+    width: 25%;
+}
+
+.author-name {
+    font-size: 15px;
+    color: #0457c8;
+    font-weight: bold;
+}
+
+.feature-ratings {
+    margin: 20px 0;
+}
+
+.feature-rating-bars {
+    margin: 10px 0;
+}
+
+.feature-header {
+    font-size: 13px;
+    font-weight: bold;
+}
+
+.review-item-body {
+    width: 75%;
+}
+
+.review-heading {
+    display: flex;
+    padding-bottom: 10px;
+}
+
+.review-title {
+    font-size: 17px;
+    display: inline-block;
+}
+
+.review-rating {
+    margin-right: 7px;
+}
+
+.verified-purchase {
+    font-weight: bold;
+    color: #1d252c;
+}
+
+.fa-check-circle {
+    padding-right: 5px;
+}
+
+.review-info-divider {
+    padding: 0 5px;
+}
+
+.post-date {
+    color: #55555a;
+}
+
+.feedback-options {
+    display: flex;
+    align-items: center;
+}
+
+.feedback-options {
+    color: #0457c8;
+}
+
+.helpful {
+    border: 1px solid #c5cbd5; 
+    padding: 0 7px;
+    margin-right: 12px;
+    height: 30px;
+    font-size: 13px;
+    background-color: #fff;
+    transition: background-color 0.2s, color 0.2s;
+}
+
+.helpful:hover {
+    background-color: #001E73;
+    color: #fff;
+}
+
+.review-content {
+    margin-top: 10px;
+    line-height: 19px;
+    font-size: 15px;
+    white-space: pre-wrap;
+}
+
+.recommendation {
+    margin-top: 20px;
+    margin-bottom: 30px;
+    display: block;
+    font-weight: bold;
+}
+
+.feedback-options {
+    display: flex;
+}
+
+.feedback-spacer {
+    height: 16px;
+    width: 1px;
+    border-color: #c5cbd5;
+    border-left: 1px solid #c5cbd5;
+    margin: 0 12px;
+    display: inline-block;
+}
+
+.stars {
+    background-image: url(../../assets/bby-stars.png);
+    background-repeat: no-repeat;
+    display: inline-block;
+    position: relative;
+    margin-bottom: 2px;
+}
+
+.stars-small {
+    width: 90px;
+    height: 18px;
+}
+
+.stars-small-5-0 {
+    background-position: -318px -54px;
+}
+
+.stars-small-4-0 {
+    background-position: -336px -54px;
+}
+
+.stars-small-3-0 {
+    background-position: -354px -54px;
+}
+
+.stars-small-2-0 {
+    background-position: -372px -54px;
+}
+
+.stars-small-1-0 {
+    background-position: -390px -54px;
+}
