+++ client/src/assets/styles_unused/VerifiedPurchaseFilter.css
@@ -1,0 +1,68 @@
+.switch-label {
+    display: flex;
+}
+
+.verified-purchases {
+    font-weight: bold;
+    color: #1d252c;
+}
+
+.switch {
+    width: 36px;
+    height: 20px;
+    position: relative;
+    display: inline-block;
+}
+
+.slider {
+    display: inline-block;
+    border-radius: 34px;
+    width: 36px;
+    height: 20px;
+    position: absolute;
+    background-color: #55555a;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    z-index: -2;
+    transition: 0.4s;
+}
+
+.slider-toggle {
+    display: inline-block;
+    border-radius: 34px;
+    width: 16px;
+    height: 16px;
+    background-color: #fff;
+    position: absolute;
+    bottom: 2px;
+    left: 2px;
+    z-index: -1;
+    transition: 0.4s;
+}
+
+.slider.checked {
+    background-color: #0046be;
+}
+
+.slider-toggle.checked {
+    left: 18px;
+}
+
+.switch > input {
+    z-index: 1;
+    width: 36px;
+    height: 20px;
+    opacity: 0;
+    cursor: pointer;
+}
+
+.slider-label {
+    margin-left: 10px;
+}
+
+.learn-more {
+    color: #0457c8;
+    margin-left: 45px;
+}
