--- client/src/components/reviews/FeatureRatingBar.css
@@ -1,28 +1,0 @@
-.rating-description {
-    display: inline-block;
-    width: 80px;
-}
-
-.rating-bar-group {
-    display: inline-flex;
-}
-
-.rating-tile {
-    height: 11px;
-    width: 22px;
-    margin: 0 1px;
-    border: 1px solid #c5cbd5;
-}
-
-.rating-tile.first {
-    border-radius: 10px 0 0 10px;
-}
-
-.rating-tile.last {
-    border-radius: 0 10px 10px 0;
-}
-
-.rating-tile.filled {
-    background-color: #55555a;
-    border: 1px solid #55555a;
-}
