--- client/src/components/reviews/RatingBar.css
@@ -1,119 +1,0 @@
-.rating-bar-container {
-    display: grid;
-    grid-template-columns: 26px 45px minmax(100px, 1fr) 60px;
-    height: 16px;
-    margin-top: 13px;
-    cursor: pointer;
-}
-
-.rating-bar-container.disabled-cursor {
-    cursor: not-allowed; 
-}
-
-.checkbox-container, .star-rating {
-    height: 16px;
-}
-
-.checkbox-container {
-    position: relative;
-    width: 16px;
-    flex: 0 0 16px;
-    margin-right: 10px;
-}
-
-.checkbox-container > input {
-    opacity: 0;
-    position: absolute;
-    height: 100%;
-    width: 100%;
-    margin: 0;
-    border: 0;
-    z-index: 1;
-    cursor: pointer;
-}
-
-.fancy-checkbox {
-    display: block;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    border-radius: 4px;
-    border-top: 1px solid #55555a;
-    border-right: 1px solid #55555a;
-    border-left: 1px solid #55555a;
-    border-bottom: 2px solid #55555a;
-    background-image: linear-gradient(135deg, #fff, #fff, 50%, #e0e6ef 0, #e0e6ef);
-}
-
-.disabled {
-    border-color: #c5cbd5;
-    background-image: none;
-    background-color: #c5cbd5;
-    border-bottom-width: 1px;
-    cursor: default;
-}
-
-.checked {
-    background-image: none
-}
-
-.checkmark {
-    display: none;
-    position: relative;
-    width: 18px;
-    height: 18px;
-    background-image: url(../../assets/checkmark.png);
-    background-repeat: no-repeat;
-    background-size: 100%;
-    top: -4px;
-    left: 1px;
-    background-position: 0px 0px;
-}
-
-.star {
-    font-size: 15px;
-    height: 16px;
-    width: 13px;
-    margin-right: 4px;
-    display: inline-block;
-}
-
-.progress-bar {
-    background-color: #e0e6ef;
-    margin-top: 3px;
-    display: inline-block;
-    height: 13px;
-    width: 100%;
-    border-radius: 6.5px;
-    box-shadow: none;
-}
-
-.progress-bar-filled {
-    float: left;
-    height: 100%;
-    background-color: #0046be;
-    line-height: 1.3;
-    border-top-left-radius: 6.5px;
-    border-bottom-left-radius: 6.5px;
-    box-shadow: inset 0 -1px 0 rgba(0,0,0,.15);
-}
-
-.gold-star{
-    background-image: url(../../assets/bby-stars.png);
-    background-repeat: no-repeat;
-    display: inline-block;
-    position: relative;
-    height: 16px;
-    width: 18px;
-    background-position: -318px -54px;
-    vertical-align: bottom;
-    margin-bottom: 1px;
-}
-
-.rating-count {
-    display: inline-block;
-    margin-left: 15px;
-    height: 16px;
-}
