--- client/src/components/reviews/RecommendationDonut.css
@@ -1,26 +1,0 @@
-.donut-container {
-    text-align: center;
-}
-
-.donut-percent {
-    position: absolute;
-    top: 36px;
-    left: 23px;
-    width: 55px;
-    font-size: 22px;
-    display: block;
-    text-align: center;
-}
-
-.donut-graphic {
-    height: 100px;
-    width: 100px;
-    display: inline-block;
-    position: relative;
-    margin: 0 auto;
-}
-
-.recommend-text {
-    margin: 0 auto;
-    text-align: center;
-}
