--- client/src/components/reviews/ReviewFilters.css
@@ -1,6 +1,0 @@
-.review-filters-container {
-    display: flex;
-    position: relative;
-    border-top: 1px solid #c5cbd5;
-    padding: 25px 0;
-}
