--- client/src/components/reviews/ReviewSummary.css
@@ -1,21 +1,0 @@
-.review-summary-container {
-    display: flex;
-    margin-top: 15px;
-    margin-bottom: 25px;
-}
-
-
-
-.spacer-column {
-    flex-grow: 1;
-    flex-shrink: 3;
-    padding: 0;
-    min-width: 10px;
-    max-width: 40px;
-}
-
-@media (min-width: 1366px) {
-    .spacer-column {
-        max-width: 60px;
-    }
-}
